% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_graph.R
\name{add_attr_to_graph}
\alias{add_attr_to_graph}
\title{Add one of a handful of graph attributes from the {tidygraph} package to your graph}
\usage{
add_attr_to_graph(
  graph,
  attr = c("centrality_degree", "group_id", "centrality_edges")
)
}
\arguments{
\item{graph}{A tidygraph object}

\item{attr}{One of 'centrality_degree', 'group_id', and 'centrality_edges'}
}
\value{
A tidygraph object with one or more added attributes
}
\description{
Add one of a handful of graph attributes from the {tidygraph} package to your graph
}
\examples{
if(FALSE){

# Add all of the available attributes.
add_attr_to_graph(my_graph)
# Or just add one / some of them.
add_attr_to_graph(my_graph, attr = c("group_id"))
}

}
