% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Copy of {bcmaps}'s ask() function}
\usage{
ask(...)
}
\arguments{
\item{...}{Unsure what this is}
}
\value{
Unsure
}
\description{
Copy of {bcmaps}'s ask() function
}
\examples{
\dontrun
}
