% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_write_to_data_dir.R
\name{check_write_to_data_dir}
\alias{check_write_to_data_dir}
\title{Copy of {bcmaps}'s same function}
\usage{
check_write_to_data_dir(dir, ask)
}
\arguments{
\item{dir}{Directory}
}
\value{
TRUE or FALSE
}
\description{
Copy of {bcmaps}'s same function
}
\examples{
\dontrun
}
