% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_away_downstream.R
\name{clip_away_downstream}
\alias{clip_away_downstream}
\title{Clip Away Downstream}
\usage{
clip_away_downstream(stream_networks, points)
}
\arguments{
\item{stream_networks}{An {sf} table describing stream network(s)}

\item{points}{An {sf} table describing one or more obstacle points}
}
\value{
The submitted stream network(s) minus their geometries downstream of the obstacle points
}
\description{
Clip Away Downstream
}
\examples{
\dontrun
}
