% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_cql_search.R
\name{compose_cql_search}
\alias{compose_cql_search}
\title{Compose CQL Search}
\usage{
compose_cql_search(fwa_codes, multi = F)
}
\value{
A single character string that encodes either a single match or a multiple match
}
\description{
Compose CQL Search
}
\examples{
\dontrun
}
