% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_stream_with_point.R
\name{cut_stream_with_point}
\alias{cut_stream_with_point}
\title{Cut Stream with Point}
\usage{
cut_stream_with_point(points, streams)
}
\arguments{
\item{points}{An {sf} table of obstacle points}

\item{streams}{As {sf} table of one or more streams}
}
\value{
Geometries of submitted stream(s) clip at the obstacle point(s) and merged by igraph component group
}
\description{
Cut Stream with Point
}
\examples{
\dontrun
}
