% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir.R
\name{data_dir}
\alias{data_dir}
\title{Thin wrapper and slight modification of {bcmaps}'s data_dir()}
\usage{
data_dir()
}
\value{
directory to store data
}
\description{
Thin wrapper and slight modification of {bcmaps}'s data_dir()
}
\examples{
\dontrun
}
