% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_comps.R
\name{delineate_comps}
\alias{delineate_comps}
\title{Delineate Component Groups}
\usage{
delineate_comps(wbs)
}
\arguments{
\item{wbs}{Stream, river, lake, etc - it must have a FWA_WATERSHED_CODE column}
}
\value{
Returns input water body table with an additional column: comp_group
}
\description{
Delineate Component Groups
}
\examples{
\dontrun
}
