% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_stream.R
\name{find_nearest_stream}
\alias{find_nearest_stream}
\title{Title Find nearest stream to a given point}
\usage{
find_nearest_stream(points, max_buffer_dist = 50)
}
\arguments{
\item{max_buffer_dist}{How far to look for stream, in meters}

\item{point}{Point to search with}
}
\value{
A single stream in an {sf} table
}
\description{
Title Find nearest stream to a given point
}
\examples{
\dontrun
}
