% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fwa_graph_obj.R
\name{fwa_graph}
\alias{fwa_graph}
\title{Title Load the FWA stream tidy graph built on the FWA watershed code field.}
\usage{
fwa_graph()
}
\value{
Large tidygraph dataset of streams in British Columbia based on Freshwater Atlas.
}
\description{
Title Load the FWA stream tidy graph built on the FWA watershed code field.
}
\examples{
fwa_graph()

}
