% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream_fwa_codes.R
\name{get_upstream_fwa_codes}
\alias{get_upstream_fwa_codes}
\title{Get a list of all FWA WATERSHED CODE'ss that are upstream of target code}
\usage{
get_upstream_fwa_codes(fwa_code)
}
\arguments{
\item{fwa_code}{Target FWA WATERSHED CODE}
}
\value{
A vector of FWA WATERSHED CODE's
}
\description{
Get a list of all FWA WATERSHED CODE'ss that are upstream of target code
}
\examples{
  fwa_pattern = '200-948755-937012-.*'
  fwa.connect::fwa_up_and_downstream_tbl[data.table::`\%like\%`(fwa.connect::fwa_up_and_downstream_tbl$upstream_fwa_code, fwa_pattern)]$upstream_fwa_code
}
