% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_invis_background_shape.R
\name{make_invis_background_shape}
\alias{make_invis_background_shape}
\title{Make Invisible Background Shape}
\usage{
make_invis_background_shape(stream, upstream_network)
}
\arguments{
\item{stream}{Obstructed stream}

\item{upstream_network}{Network of streams upstream from obstruction}
}
\value{
A bounding box of equal height and width, the length of which is the longest dimension of the plotted objects
}
\description{
Make Invisible Background Shape
}
\examples{
\dontrun
}
