% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_conn_tbl.R
\name{stream_conn_tbl}
\alias{stream_conn_tbl}
\title{Download and load in the FWA stream connectivity table.}
\usage{
stream_conn_tbl(ask = interactive(), force = FALSE)
}
\value{
Table showing connectivity of streams in British Columbia based on the Freshwater Atlas Stream Network.
}
\description{
Download and load in the FWA stream connectivity table.
}
\examples{
str_tbl = stream_conn_tbl()
str_tbl

}
