% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_course_upstream.R
\name{trace_course_upstream}
\alias{trace_course_upstream}
\title{Title Get all streams (and their geometries) upstream of a given FWA code.}
\usage{
trace_course_upstream(
  fwa_code,
  merge_by_BLK = T,
  make_plot = F,
  add_map_insert = F,
  save_plot = F,
  save_plot_location = NULL
)
}
\arguments{
\item{fwa_code}{The FWA_WATERSHED_CODE of a target stream}

\item{merge_by_BLK}{Merge resulting stream table by BLUE_LINE_KEY?}

\item{make_plot}{Make ggplot of results?}

\item{add_map_insert}{Add a little insert of BC to give spatial context of main plot?}

\item{save_plot}{Save .PNG file of plot to local machine?}

\item{save_plot_location}{Where to save plot; defaults to current working directory}
}
\value{
An {sf} spatial table of stream with submitted FWA code plus all upstream streams
}
\description{
Title Get all streams (and their geometries) upstream of a given FWA code.
}
\examples{
if(FALSE){
paul_river_upstream = trace_course_upstream(fwa_code = "200-948755-999851-274772-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000",
make_plot = T, add_map_insert = T)
paul_river_upstream
}
}
