% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_stream_with_point.R
\name{cut_stream_with_point}
\alias{cut_stream_with_point}
\title{Cut Stream with Point}
\usage{
cut_stream_with_point(point, obstructed_stream, buffer_point_amount = 50)
}
\arguments{
\item{point}{Barrier (point)}

\item{obstructed_stream}{Sf geometry of an obstructed stream}

\item{buffer_point_amount}{How many meters should the obstruction point be buffered by?}
}
\value{
Returns two lines from one original stream line, cut by the point
}
\description{
Cut Stream with Point
}
\examples{
\dontrun
}
