% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream_fwa_codes.R
\name{get_upstream_fwa_codes}
\alias{get_upstream_fwa_codes}
\title{Get a list of all FWA WATERSHED CODEs that are upstream of target code}
\usage{
get_upstream_fwa_codes(fwa_code)
}
\arguments{
\item{fwa_code}{Target FWA WATERSHED CODE}
}
\value{
A vector of FWA WATERSHED CODEs
}
\description{
Get a list of all FWA WATERSHED CODEs that are upstream of target code
}
\examples{
\dontrun
}
